Contract class extraction
